u-boot: workaround for setlocalversion issue
The u-boot build process attempts to determine the version number using git
commands, which assumes that the build is taking place in a git checkout with
the .git directory intact. The buildroot build process does a git checkout, but
then exports a .tar.gz as a cache and does the build from that tarball (with no
.git file).

In some instances, being unable to determine the current version causes the
build process to generate a non-functional bootloader.

This patch implements a work-around for this problem by allowing the user to
configure a 'localverion' override which gets written to the ./localversion
file in the u-boot build directory, which is later used by the
./tools/setlocalversion script.